(chore): implements oauth2proxy authentication via dex
Service Proxy ingress uses OAuth2Proxy to authenticate against Dex
