+++ internal/controller/organization/organization_controller.go
@@ -12,7 +12,6 @@ import (
 	"github.com/pkg/errors"
 	corev1 "k8s.io/api/core/v1"
 	rbacv1 "k8s.io/api/rbac/v1"
-	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/tools/record"
 	ctrl "sigs.k8s.io/controller-runtime"
@@ -50,9 +49,6 @@ var (
 
 const (
 	defaultGreenhouseConnectorID = "greenhouse"
-	// internalSuffix is used for the internal secret of the organization
-	// this secret is used to store secrets that are not created by the user
-	technicalSecretSuffix = "-internal"
 )
 
 // OrganizationReconciler reconciles an Organization object
@@ -181,10 +177,6 @@ func (r *OrganizationReconciler) EnsureCreated(ctx context.Context, object lifec
 				org.SetCondition(greenhousev1alpha1.FalseCondition(greenhousev1alpha1.OrganizationOICDConfigured, greenhousev1alpha1.DefaultConnectorRedirectsFailed, err.Error()))
 				return ctrl.Result{}, lifecycle.Failed, err
 			}
-		}
-		if err := r.reconcileOAuth2ProxySecret(ctx, org); err != nil {
-			org.SetCondition(greenhousev1alpha1.FalseCondition(greenhousev1alpha1.OrganizationOICDConfigured, greenhousev1alpha1.OAuthProxySecretFailed, err.Error()))
-			return ctrl.Result{}, lifecycle.Failed, err
 		}
 		org.SetCondition(greenhousev1alpha1.TrueCondition(greenhousev1alpha1.OrganizationOICDConfigured, "", ""))
 	}
@@ -373,36 +365,4 @@ func (r *OrganizationReconciler) enqueueOrganizationForReferencedSecret(_ contex
 		return nil
 	}
 	return []ctrl.Request{{NamespacedName: client.ObjectKeyFromObject(org)}}
-}
-
-// getOrCreateOrgSecret creates the internal secret of an organization, used to store secrets that are not created by the user.
-// The secret is created with the name <org.Name>-internal and the namespace <org.Namespace>.
-func (r *OrganizationReconciler) getOrCreateOrgSecret(ctx context.Context, org *greenhousev1alpha1.Organization) (*corev1.Secret, error) {
-	secret := new(corev1.Secret)
-	secret.Name = org.Name + technicalSecretSuffix
-	secret.Namespace = org.Name
-	secret.Type = greenhouseapis.SecretTypeOrganization
-
-	// check if the secret already exists
-	err := r.Get(ctx, types.NamespacedName{Namespace: secret.Namespace, Name: secret.Name}, secret)
-	if err == nil {
-		return secret, nil
-	}
-	if !apierrors.IsNotFound(err) {
-		return nil, err
-	}
-
-	result, err := clientutil.CreateOrPatch(ctx, r.Client, secret, func() error {
-		return controllerutil.SetControllerReference(org, secret, r.Scheme())
-	})
-	if err != nil {
-		return nil, err
-	}
-	switch result {
-	case clientutil.OperationResultCreated:
-		log.FromContext(ctx).Info("created secret", "name", secret.Name)
-	case clientutil.OperationResultUpdated:
-		log.FromContext(ctx).Info("updated secret", "name", secret.Name)
-	}
-	return secret, nil
 }
